Struct aws_sdk_transcribestreaming::client::fluent_builders::StartStreamTranscription

pub struct StartStreamTranscription<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartStreamTranscription<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<StartStreamTranscriptionOutput, SdkError<StartStreamTranscriptionError>> where
R::Policy: SmithyRetryPolicy<StartStreamTranscriptionInputOperationOutputAlias, StartStreamTranscriptionOutput, StartStreamTranscriptionError, StartStreamTranscriptionInputOperationRetryAlias>,

pub fn language_code(self, inp: LanguageCode) -> Self

Indicates the source language used in the input audio stream.

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn media_sample_rate_hertz(self, inp: i32) -> Self

The sample rate, in Hertz, of the input audio. We suggest that you use 8000 Hz for low quality audio and 16000 Hz for high quality audio.

pub fn set_media_sample_rate_hertz(self, input: Option<i32>) -> Self

pub fn media_encoding(self, inp: MediaEncoding) -> Self

The encoding used for the input audio.

pub fn set_media_encoding(self, input: Option<MediaEncoding>) -> Self

pub fn vocabulary_name(self, inp: impl Into<String>) -> Self

The name of the vocabulary to use when processing the transcription job.

pub fn set_vocabulary_name(self, input: Option<String>) -> Self

pub fn session_id(self, inp: impl Into<String>) -> Self

A identifier for the transcription session. Use this parameter when you want to retry a session. If you don't provide a session ID, Amazon Transcribe will generate one for you and return it in the response.

pub fn set_session_id(self, input: Option<String>) -> Self

pub fn audio_stream(self, inp: EventStreamInput<AudioStream>) -> Self

PCM-encoded stream of audio blobs. The audio stream is encoded as an HTTP2 data frame.

pub fn set_audio_stream(
self,
input: Option<EventStreamInput<AudioStream>>
) -> Self

pub fn vocabulary_filter_name(self, inp: impl Into<String>) -> Self

The name of the vocabulary filter you've created that is unique to your AWS account. Provide the name in this field to successfully use it in a stream.

pub fn set_vocabulary_filter_name(self, input: Option<String>) -> Self

pub fn vocabulary_filter_method(self, inp: VocabularyFilterMethod) -> Self

The manner in which you use your vocabulary filter to filter words in your transcript. Remove removes filtered words from your transcription results. Mask masks those words with a *** in your transcription results. Tag keeps the filtered words in your transcription results and tags them. The tag appears as VocabularyFilterMatch equal to True

pub fn set_vocabulary_filter_method(
self,
input: Option<VocabularyFilterMethod>
) -> Self

pub fn show_speaker_label(self, inp: bool) -> Self

When true, enables speaker identification in your real-time stream.

pub fn set_show_speaker_label(self, input: Option<bool>) -> Self

pub fn enable_channel_identification(self, inp: bool) -> Self

When true, instructs Amazon Transcribe to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe also produces a transcription of each item. An item includes the start time, end time, and any alternative transcriptions.

You can't set both ShowSpeakerLabel and EnableChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

pub fn set_enable_channel_identification(self, input: Option<bool>) -> Self

pub fn number_of_channels(self, inp: i32) -> Self

The number of channels that are in your audio stream.

pub fn set_number_of_channels(self, input: Option<i32>) -> Self

pub fn enable_partial_results_stabilization(self, inp: bool) -> Self

When true, instructs Amazon Transcribe to present transcription results that have the partial results stabilized. Normally, any word or phrase from one partial result can change in a subsequent partial result. With partial results stabilization enabled, only the last few words of one partial result can change in another partial result.

pub fn set_enable_partial_results_stabilization(
self,
input: Option<bool>
) -> Self

pub fn partial_results_stability(self, inp: PartialResultsStability) -> Self

You can use this field to set the stability level of the transcription results. A higher stability level means that the transcription results are less likely to change. Higher stability levels can come with lower overall transcription accuracy.

pub fn set_partial_results_stability(
self,
input: Option<PartialResultsStability>
) -> Self