Module aws_sdk_transcribestreaming::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

service_unavailable_exception

See ServiceUnavailableException

Structs

BadRequestException

One or more arguments to the StartStreamTranscription or

ConflictException

A new stream started with the same session ID. The current stream has been

InternalFailureException

A problem occurred while processing the audio. Amazon Transcribe or Amazon Transcribe Medical terminated processing. Try

LimitExceededException

You have exceeded the maximum number of concurrent transcription streams, are starting

ServiceUnavailableException

Service is currently unavailable. Try your request later.

StartMedicalStreamTranscriptionError

StartStreamTranscriptionError

Enums

StartMedicalStreamTranscriptionErrorKind

StartStreamTranscriptionErrorKind