Module aws_sdk_transcribestreaming::model

Modules

alternative

See Alternative

audio_event

See AudioEvent

item

See Item

medical_alternative

See MedicalAlternative

medical_entity

See MedicalEntity

medical_item

See MedicalItem

medical_result

See MedicalResult

medical_transcript

See MedicalTranscript

medical_transcript_event

See MedicalTranscriptEvent

result

See Result

transcript

See Transcript

transcript_event

See TranscriptEvent

Structs

Alternative

A list of possible transcriptions for the audio.

AudioEvent

Provides a wrapper for the audio chunks that you are sending.

Item

A word, phrase, or punctuation mark that is transcribed from the input audio.

MedicalAlternative

A list of possible transcriptions for the audio.

MedicalEntity

The medical entity identified as personal health information.

MedicalItem

A word, phrase, or punctuation mark that is transcribed from the input audio.

MedicalResult

The results of transcribing a portion of the input audio stream.

MedicalTranscript

The medical transcript in a MedicalTranscriptEvent.

MedicalTranscriptEvent

Represents a set of transcription results from the server to the client. It contains

Result

The result of transcribing a portion of the input audio stream.

Transcript

The transcription in a TranscriptEvent.

TranscriptEvent

Represents a set of transcription results from the server to the client. It contains one

Enums

AudioStream

ItemType

LanguageCode

MediaEncoding

MedicalContentIdentificationType

MedicalTranscriptResultStream

PartialResultsStability

Specialty

TranscriptResultStream

Type

VocabularyFilterMethod