Struct aws_sdk_transcribestreaming::output::start_medical_stream_transcription_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for StartMedicalStreamTranscriptionOutput

Implementations

impl Builder

pub fn request_id(self, input: impl Into<String>) -> Self

An identifier for the streaming transcription.

pub fn set_request_id(self, input: Option<String>) -> Self

pub fn language_code(self, input: LanguageCode) -> Self

The language code for the response transcript. For Amazon Transcribe Medical, this is US English (en-US).

pub fn set_language_code(self, input: Option<LanguageCode>) -> Self

pub fn media_sample_rate_hertz(self, input: i32) -> Self

The sample rate of the input audio in Hertz. Valid value: 16000 Hz.

pub fn set_media_sample_rate_hertz(self, input: Option<i32>) -> Self

pub fn media_encoding(self, input: MediaEncoding) -> Self

The encoding used for the input audio stream.

pub fn set_media_encoding(self, input: Option<MediaEncoding>) -> Self

pub fn vocabulary_name(self, input: impl Into<String>) -> Self

The name of the vocabulary used when processing the stream.

pub fn set_vocabulary_name(self, input: Option<String>) -> Self

pub fn specialty(self, input: Specialty) -> Self

The specialty in the medical domain.

pub fn set_specialty(self, input: Option<Specialty>) -> Self

pub fn type(self, input: Type) -> Self

The type of audio that was transcribed.

pub fn set_type(self, input: Option<Type>) -> Self

pub fn show_speaker_label(self, input: bool) -> Self

Shows whether speaker identification was enabled in the stream.

pub fn set_show_speaker_label(self, input: Option<bool>) -> Self

pub fn session_id(self, input: impl Into<String>) -> Self

Optional. An identifier for the transcription session. If you don't provide a session ID, Amazon Transcribe generates one for you and returns it in the response.

pub fn set_session_id(self, input: Option<String>) -> Self

pub fn transcript_result_stream(
self,
input: Receiver<MedicalTranscriptResultStream, StartMedicalStreamTranscriptionError>
) -> Self

Represents the stream of transcription events from Amazon Transcribe Medical to your application.

pub fn set_transcript_result_stream(
self,
input: Option<Receiver<MedicalTranscriptResultStream, StartMedicalStreamTranscriptionError>>
) -> Self

pub fn enable_channel_identification(self, input: bool) -> Self

Shows whether channel identification has been enabled in the stream.

pub fn set_enable_channel_identification(self, input: Option<bool>) -> Self

pub fn number_of_channels(self, input: i32) -> Self

The number of channels identified in the stream.

pub fn set_number_of_channels(self, input: Option<i32>) -> Self

pub fn content_identification_type(
self,
input: MedicalContentIdentificationType
) -> Self

If the value is PHI, indicates that you've configured your stream to identify personal health information.

pub fn set_content_identification_type(
self,
input: Option<MedicalContentIdentificationType>
) -> Self

pub fn build(self) -> Result<StartMedicalStreamTranscriptionOutput, BuildError>

Consumes the builder and constructs a StartMedicalStreamTranscriptionOutput