Struct aws_sdk_transcribestreaming::output::StartStreamTranscriptionOutput

#[non_exhaustive]
pub struct StartStreamTranscriptionOutput {

    pub request_id: Option<String>,
    pub language_code: Option<LanguageCode>,
    pub media_sample_rate_hertz: Option<i32>,
    pub media_encoding: Option<MediaEncoding>,
    pub vocabulary_name: Option<String>,
    pub session_id: Option<String>,
    pub transcript_result_stream: Receiver<TranscriptResultStream, StartStreamTranscriptionError>,
    pub vocabulary_filter_name: Option<String>,
    pub vocabulary_filter_method: Option<VocabularyFilterMethod>,
    pub show_speaker_label: bool,
    pub enable_channel_identification: bool,
    pub number_of_channels: Option<i32>,
    pub enable_partial_results_stabilization: bool,
    pub partial_results_stability: Option<PartialResultsStability>,
    pub content_identification_type: Option<ContentIdentificationType>,
    pub content_redaction_type: Option<ContentRedactionType>,
    pub pii_entity_types: Option<String>,
    pub language_model_name: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

request_id: Option<String>

An identifier for the streaming transcription.

language_code: Option<LanguageCode>

The language code for the input audio stream.

media_sample_rate_hertz: Option<i32>

The sample rate for the input audio stream. Use 8,000 Hz for low quality audio and 16,000 Hz for high quality audio.

media_encoding: Option<MediaEncoding>

The encoding used for the input audio stream.

vocabulary_name: Option<String>

The name of the vocabulary used when processing the stream.

session_id: Option<String>

An identifier for a specific transcription session.

transcript_result_stream: Receiver<TranscriptResultStream, StartStreamTranscriptionError>

Represents the stream of transcription events from Amazon Transcribe to your application.

vocabulary_filter_name: Option<String>

The name of the vocabulary filter used in your real-time stream.

vocabulary_filter_method: Option<VocabularyFilterMethod>

The vocabulary filtering method used in the real-time stream.

show_speaker_label: bool

Shows whether speaker identification was enabled in the stream.

enable_channel_identification: bool

Shows whether channel identification has been enabled in the stream.

number_of_channels: Option<i32>

The number of channels identified in the stream.

enable_partial_results_stabilization: bool

Shows whether partial results stabilization has been enabled in the stream.

partial_results_stability: Option<PartialResultsStability>

If partial results stabilization has been enabled in the stream, shows the stability level.

content_identification_type: Option<ContentIdentificationType>

Shows whether content identification was enabled in this stream.

content_redaction_type: Option<ContentRedactionType>

Shows whether content redaction was enabled in this stream.

pii_entity_types: Option<String>

Lists the PII entity types you specified in your request.

language_model_name: Option<String>
Implementations

impl StartStreamTranscriptionOutput

pub fn request_id(&self) -> Option<&str>

An identifier for the streaming transcription.

pub fn language_code(&self) -> Option<&LanguageCode>

The language code for the input audio stream.

pub fn media_sample_rate_hertz(&self) -> Option<i32>

The sample rate for the input audio stream. Use 8,000 Hz for low quality audio and 16,000 Hz for high quality audio.

pub fn media_encoding(&self) -> Option<&MediaEncoding>

The encoding used for the input audio stream.

pub fn vocabulary_name(&self) -> Option<&str>

The name of the vocabulary used when processing the stream.

pub fn session_id(&self) -> Option<&str>

An identifier for a specific transcription session.

pub fn transcript_result_stream(
&self
) -> &Receiver<TranscriptResultStream, StartStreamTranscriptionError>

Represents the stream of transcription events from Amazon Transcribe to your application.

pub fn vocabulary_filter_name(&self) -> Option<&str>

The name of the vocabulary filter used in your real-time stream.

pub fn vocabulary_filter_method(&self) -> Option<&VocabularyFilterMethod>

The vocabulary filtering method used in the real-time stream.

pub fn show_speaker_label(&self) -> bool

Shows whether speaker identification was enabled in the stream.

pub fn enable_channel_identification(&self) -> bool

Shows whether channel identification has been enabled in the stream.

pub fn number_of_channels(&self) -> Option<i32>

The number of channels identified in the stream.

pub fn enable_partial_results_stabilization(&self) -> bool

Shows whether partial results stabilization has been enabled in the stream.

pub fn partial_results_stability(&self) -> Option<&PartialResultsStability>

If partial results stabilization has been enabled in the stream, shows the stability level.

pub fn content_identification_type(&self) -> Option<&ContentIdentificationType>

Shows whether content identification was enabled in this stream.

pub fn content_redaction_type(&self) -> Option<&ContentRedactionType>

Shows whether content redaction was enabled in this stream.

pub fn pii_entity_types(&self) -> Option<&str>

Lists the PII entity types you specified in your request.

pub fn language_model_name(&self) -> Option<&str>

impl StartStreamTranscriptionOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture StartStreamTranscriptionOutput